requirements.txt
fastapi
uvicorn[standard]
pydantic
starlette
sqlalchemy
bcrypt==4.0.1
passlib[bcrypt]
python-multipart
python-jose[crytography]
alembic
pytest
httpx
pytest-asyncio